Added reverse shooter, lower port mode, and more

+++ src/main/java/org/longmetal/subsystem/DriveTrain.java
@@ -74,7 +74,7 @@ public class DriveTrain {
         driveTrain.curvatureDrive(driveX, driveZ, true); // Drive
     }
 
-    public void curveRaw(int xSpeed, double zRotation, boolean isQuickTurn) {
+    public void curveRaw(double xSpeed, double zRotation, boolean isQuickTurn) {
         driveTrain.curvatureDrive(xSpeed, zRotation, isQuickTurn);
     }
 }
